requirements.txt
flask>=2.0
exifread>=3.0
pillow>=9.0
python-dotenv>=0.19
requests>=2.28
pyinstaller>=6.0
pywebview>=5.0
